Move CompletionDrawSupport to parent package.
This way I can reduce the visibility of the constructor in
TableOwnerDrawSupport to default and avoid exposing more (internal) API
surface than strictly necessary.
